[general.ios] did #60: now also compiling on arm64, compiler warning unveiled some implicit integer conversions now made explicit. Still many in thrift auto-generated files though.

+++ ios/PocketCampus/Plugins/MapPlugin/Model/MapUtils.m
@@ -312,7 +312,7 @@ double WGStoCHy(double lat, double lng) {
 }
 
 + (MKCoordinateRegion)regionToFitMapItemAnnotations:(NSArray*)annotations {
-    int count = annotations.count;
+    NSUInteger count = annotations.count;
     if (count == 0) {
         return MKCoordinateRegionMake(CLLocationCoordinate2DMake(0, 0), MKCoordinateSpanMake(0, 0));
     }
